test 10

+++ build.sh
@@ -5,7 +5,7 @@ git clone --branch v6.8.3 https://github.com/qt/qtbase.git qt_lts
 mkdir qt_build
 cd qt_build
 ../qt_lts/configure -h
-../qt_lts/configure -submodules qtbase,qtnetwork,qtserialport -no-sbom
+../qt_lts/configure -submodules qtbase,qtnetwork,qtserialport -no-sbom -no-dbus -no-gui -no-widgets
 if [ "$?" -ne "0" ]; then
   echo "Qt configuration failed"
   exit 1
@@ -16,6 +16,10 @@ if [ "$?" -ne "0" ]; then
   exit 1
 fi
 cmake --install .
+if [ "$?" -ne "0" ]; then
+  echo "Qt install failed"
+  exit 1
+fi
 
 cd ..
 rm -r qt_lts
